+++ src/index.ts
@@ -1,5 +1,6 @@
 import enforcePlaceholders from './rules/enforce-placeholders.js'
 import noDynamicKey from './rules/no-dynamic-key.js'
+import noUnusedPlaceholderParams from './rules/no-unused-placeholder-params.js'
 import preferStringArguments from './rules/prefer-string-arguments.js'
 import requireKey from './rules/require-key.js'
 
@@ -13,6 +14,7 @@ const plugin = {
     'prefer-string-arguments': preferStringArguments,
     'no-dynamic-key': noDynamicKey,
     'enforce-placeholders': enforcePlaceholders,
+    'no-unused-placeholder-params': noUnusedPlaceholderParams,
   },
   configs: {} as Record<string, unknown>,
 }
@@ -26,6 +28,7 @@ plugin.configs.recommended = {
     'tolgee/prefer-string-arguments': 'error',
     'tolgee/no-dynamic-key': 'error',
     'tolgee/enforce-placeholders': 'error',
+    'tolgee/no-unused-placeholder-params': 'error',
   },
 }